mv handleConstraintDefaults into constraints.js file
